+++ package.json
@@ -127,6 +127,7 @@
     "react-native-code-push": "^7.0.4",
     "react-native-country-picker-modal": "^2.0.0",
     "react-native-dash": "^0.0.11",
+    "react-native-date-picker": "^4.1.3",
     "react-native-device-info": "^8.1.0",
     "react-native-email-link": "^1.10.0",
     "react-native-fast-image": "^8.3.4",
