+++ src/features/auth/signup/SetBirthday/BirthdayInformationModal/BirthdayInformationModal.tsx
